+++ DTCD-WorkspaceSystem/src/WorkspaceSystem.js
@@ -1,6 +1,7 @@
 import './styles/panel.css';
 import './styles/header.css';
 import './styles/footer.css';
+import './styles/modal.css';
 
 import {
   EventSystemAdapter,
@@ -47,6 +48,7 @@ export class WorkspaceSystem extends SystemPlugin {
   #grid;
   #editMode;
   #numberPanelIncrement;
+  #modalInstance;
 
   static getRegistrationMeta() {
     return {
@@ -70,6 +72,7 @@ export class WorkspaceSystem extends SystemPlugin {
 
     this.#panels = [];
     this.#editMode = false;
+    this.#modalInstance = null;
 
     toMountTemplates();
 
@@ -161,7 +164,7 @@ export class WorkspaceSystem extends SystemPlugin {
 
     // ---- PLUGINS ----
 
-    let subscriptions; // From workspace config for eventSystem process only subscriptions
+    let eventSystemConfig = {};
 
     // ---- installing-plugins-from-config ----
     const GUIDMap = {};
@@ -193,7 +196,7 @@ export class WorkspaceSystem extends SystemPlugin {
           GUIDMap[guid] = systemGUID;
 
           if (meta.name === 'EventSystem') {
-            subscriptions = config.subscriptions;
+            eventSystemConfig = config;
             continue pluginsLoop;
           }
           if (meta.name === 'WorkspaceSystem') continue pluginsLoop;
@@ -208,14 +211,14 @@ export class WorkspaceSystem extends SystemPlugin {
     }
 
     // EVENT-SYSTEM-MAPPING
-    if (subscriptions)
-      for (let sub of subscriptions) {
+    if (eventSystemConfig.hasOwnProperty('subscriptions'))
+      for (let sub of eventSystemConfig.subscriptions) {
         const { event, action } = sub;
         event.guid = GUIDMap[event.guid];
         action.guid = GUIDMap[action.guid];
       }
 
-    await this.getSystem('EventSystem').setPluginConfig({ subscriptions });
+    await this.getSystem('EventSystem').setPluginConfig(eventSystemConfig);
     return true;
   }
 
@@ -542,5 +545,46 @@ export class WorkspaceSystem extends SystemPlugin {
     styleEl.innerHTML = style;
     head.appendChild(styleEl);
     this.#column = column;
+  }
+
+  openPanelInModal(panelName) {
+    if (!this.#modalInstance) {
+      const modalBackdrop = document.createElement('div');
+      modalBackdrop.classList.add('modal-backdrop');
+      modalBackdrop.id = 'modal-backdrop';
+      const modal = document.createElement('div');
+      modal.classList.add('modal');
+
+      const panelContainer = document.createElement('div');
+      panelContainer.id = 'mount-point';
+      modal.appendChild(panelContainer);
+      modalBackdrop.appendChild(modal);
+
+      modalBackdrop.addEventListener('click', evt => {
+        if (evt.target.isEqualNode(modalBackdrop)) {
+          this.closeModal();
+        }
+      });
+
+      modal.addEventListener('click', evt => {
+        evt.stopPropagation();
+      });
+
+      try {
+        const plugin = this.getPlugin(panelName);
+        document.body.append(modalBackdrop);
+        this.#modalInstance = new plugin('', '#mount-point', true);
+      } catch (err) {
+        this.#logSystem.error(`Can't create modal with panel '${panelName}' due to error: ${err}`);
+      }
+    }
+  }
+
+  closeModal() {
+    const modal = document.getElementById('modal-backdrop');
+    if (modal) {
+      modal.remove();
+      this.#modalInstance = null;
+    }
   }
 }